requirements.txt
pandas
matplotlib
openai
python-dotenv
azure-core
azure-identity
numpy
pillow
openpyxl


